DEV: Move javascript code in script tag to theme javascript file
Why this change?

This is mostly to improve the dev experience but also for performance as
well since theme javascript files can be served via the CDN whereas
javascript defined within a `<script>` tag has to be shipped to the
client on every full page load leading to a larger initial response body.